add type cast in "ccp" marco, to avoid warnings

+++ cocos2dx/include/CGPointExtension.h
@@ -50,7 +50,7 @@ namespace   cocos2d {
  @return CGPoint
  @since v0.7.2
  */
-#define ccp(__X__,__Y__) CGPointMake(__X__,__Y__)
+#define ccp(__X__,__Y__) CGPointMake((float)__X__, (float)__Y__)
 
 /** Returns opposite of point.
  @return CGPoint
